Remove FailFatal to match servant-server’s behavior
Previously miso made a distinction between fatal errors caused by
failures to parse parameters and non-fatal errors. In the case of
fatal errors miso did not try other routes. However, this does not
match the behavior of servant-server which will try other routes on
parse failures. Since the same routes are used on the server and the
client this can cause problems (and it is also simply confusing for
people coming from servant-server).