Introduce LocatedView to make the location (URI) part of the View
Conceptually the location should be part of the view since it's what
the user sees in her browser. Currently changing the URI means going
via IO (pushURI). This is unfortunate since miso is about having a
pure view function.

This patch changes the type of the view function to return a:

  data LocatedView action = LocatedView
    { lvLocation :: Maybe URI
    , lvView :: View action
    }

When the view function returns a LocatedView with lvLocation set to
`Just uri` and `uri` is different than the current URI, `uri` will be
pushed on the history using pushURI.