Introduce 1ms delay during unmount.
In scenarios (like the components example) where deeply nested
components undergo unmounting in rapid succession, a delay must be introduced to ensure
that unmount actions are processed by the parent before the thread is
killed. Otherwise, this can lead to scenarios where the middle child
does not have enough time to process both the unmount action from its
own child and the relaying of its unmount action to its parent.