Add Miso.Lens module, re-export.
Miso applications consist of very simple state management in a State
monad. Given the shortcomings of Haskell records it is incumbent upon
us to provide users with a simple way to get / set their records.

This module is meant to make miso more "batteries-included" and reduce
dependency footprint, along  with exposing new users to lensing, but
with a seatbelt. This lens formulation is very simple (much simpler than
microlens and the original 'lens' library). It exposes an
identical interface wih fixity parity of existing combinators in microlens.

Some combinators have been omitted for brevity, others because of
conflicts (e.g. @view@ is omitted because of conflicts, and @Effect@
only operates in 'MonadState'). We are attempting
to make an identical interface (so as not to confuse LLMs) and to
make it easy for users to upgrade to more mature, sophisticated lens
libraries at their leisure.

- Update README.md and examples/simple/Main.hs to use Miso.Lens
- Document combinators, smart constructors, haddockage