Determine child reconciliation statically. (#881)
* Determine child reconciliation statically.

The child reconciliation algorithm is entered during dynamic detection
of keyed VNode. This is inefficient during the diff, and can also be
determined at tree construction time.

This PR does the following:

- Introduces the 'shouldSync' property in types.ts
- Adds 'shouldSync' FFI call in the Haskell layer
- Updates 'runView' in the 'VNode' case to set shouldSync during tree construction
- Removes code related to dynamic checking ('hasKeys')
- Modifies 'diffChildren' to accept 'VNode' rather than 'Array<VNode>'
- Defaults 'shouldSync' to 'false' in smart.ts
- Encapsulates 'shouldSync' in 'vnode()' (ensures test logic is correct)
- Regenerates JavaScripts

* Re-export for Haskell layer.

* Regen JS.

* lint

* If shouldSync is already set then bypass linear time check.

This will be a performance gain, since we won't have to traverse the
entire child list. Turns a linear time check to constant time, at the
cost of some correctness.

* Re-export shouldSync onto globalThis

* Eval shouldSync, re-export, regen JS