Simplify `Component`.
- Consolidates 'embed' and 'component' into the same function
- `SomeComponent` -> `SomeApp`
- Drop `Component` data type altogether
- `notify` and `sample` now take `MisoString`
- Introduce `componentWith`, `componentWith_`, `component`, `component_`
- Drop dead `Component`-related code

Since the introduction of dynamically generated components it is now
possible for a parent `App` to get the component name of its child using `ask`.

While this is less type-safe, it supports the dynamic `Component`
workflow better. We might consider parameterizing `App` by `Symbol` to
recover the lost type safety, `SomeSymbol` could be used as well for
dynamic `Component` generation.