Add Miso.Util.Parser (#973)
* Add Miso.Util.Lexer

Adds a lexer that operates on MisoString in Utils/ submodules.

- [x] Adds Lexer type that operates on `MisoString`
- [x] Adds combinators for `char` and `string` (using `MisoString`)

This will be useful for creating lexers, parsers, and QQs, that
operate on HTML (jsx-esque), CSS, URIs.

* Add Miso.Util.Parser

Builds on `Miso.Util.Lexer` in #972

Adds a parser that operates on `[token]` in `Utils/` submodules. This
is based on the classic Hutton Monadic Parser combinator formulation.

- [x] Adds `Parser` `Monad`
- [x] Adds combinators for parsing, `satisfy`, `sepBy`, `peek`, `oneOf`
- [x] Adjusts some shadowing in `Miso.Util.Lexer`

This will be useful for creating lexers, parsers, and QQs, that operate on HTML (jsx-esque), CSS, URIs.