Add `modelCheck :: Bool` to `Component`.
`modelCheck` is used to decide if we should diff the model before diffing the virtual DOM.
By default this is 'True'. It is recommended to set to 'False' when
developing full-screen 'canvas' applications. If 'False' model diffing
is bypassed and the 'Component' will not block on an empty event queue.

This should make canvas applications slightly more performant.