Simplify `componentId`.
With the advent of https://github.com/dmjio/miso/pull/1002, Component
communication is now indirect via the pub / sub system. Therefore
Component are no longer directly addressable via 'data-component-id',
and all Component are now dynamic. A lot of code related to
'data-component-id' can be removed, including the slow
`document.querySelectorAll` to detect duplicate Component mounting.

This PR drops explicit setting of 'data-component-id' on Component
HTML ('component-id' exists, but on the virtual DOM only).

Furthermore, previously component mounting relied on
`data-component-id`, we can now pass the `DOMRef` in explicitly from
the JS side in the call to `mount()` (along with a callback),
and avoid an expensive lookup in doing so.

- [x] Drops `Miso.Exception` (duplicate mounting is now impossible)
- [x] Drops `setComponentId`
- [x] Update Component tests (rework tests to avoid `'data-component-id'`, compare on `nodeType`)
- [x] `freshComponentId` no longer needs to be used top-level in `miso` nor `startComponent`
- [x] Adds `syncCallback2` to address mounting callback (passes in `DOMRef` explicitly)
- [x] Removes `Miso.Exception` (DuplicateMounting is no longer possible)
- [x] Updates comments to reflect the reality of Component diffing in `diffNodes`.
- [x] `data-component-id` -> `component-id`