Adds `.parentNode` check during delegation.
Each event delegator runs separately for each component. How event
handling occurs is dependent on the scheduler and the synchronous
callback implementation, this varies by backend (GHC / WASM / GHCJS).

In WASM, during component unmounting, event undelegation is not
occuring the same way as the JS backend (due to a thread scheduling
difference). Events raised from freshly unmounted components can reach
`document` (this does not occur in GHCJS86/9122). This has the effect
of events reaching `document` and `document.parentNode` is `null`.

This patch avoids that from occuring for the WASM backend.