📬 Adds `Component` `mailbox`
This PR adds a `Mailbox` to each `Component`. This means any
`Component` can send any `Value` to any other `Component` (including
itself). `CompenentId` can be obtained via on `onMountedWith`
callback, or published to a `Topic` for distribution via the Pub / Sub
mechanism (this can be done using the 'onMounted' hook as well).

- [x] Adds 'mail' function, used for messaging
- [x] Adds 'getComponentId', 'getParentComponentId' helper functions in FFI and in Effect
- [x] Adds an existential 'mailbox' to 'Component', defaults to `\_ -> Nothing`