Add helper functions `optionalAttrs`, `optionalChildren`. (#1102)
This is to avoid situations where list concatenation in the `View`
becomes awkward in the presence of conditional logic. For example,

```haskell
view m =
  div_
  ([ id_ "some-id"
   ] ++
   [ className "foo" | m ^. hasFoo
   ])
   [ "test" ]
```

This is a common pattern, and it can now be simplified to the below

```haskell
view m =
  optionalAttrs
  div_
   [ id_ "some-id"
   ]
   (m ^. hasFoo)
   [ className "foo"]
   [ "test" ]
```

The above example solves this problem for `Attribute`, but the same
exists with child lists, as seen below.

```haskell
view m =
  div_
   [ id_ "some-id"
   ]
   [ "test" ]
   (m ^. hasFoo)
   [ "foo" ]
```

In the above, the conditional `m ^. hasFoo` will * append * the child
list `[ "foo" ]` to `[ "test" ]` resulting in `[ "foo", "test" ]`.