Add `WebSocket` `connect` variants for each Content-Type.
Currently the `connect` function in `Miso.WebSocket` has a `Payload`
return type. This allows the client to receive any valid type of
message from the server. `miso` will identify if the content type
matches raw text, json, array buffer, or blob and dispatch on the
appropriate callback function.

The problem with this is that `JSON.parse` will correctly parse text
information as JSON (e.g. `"null"`, `"true"`, `3`, etc.). It is not
possible for us to identify if user intent is to receive /only/ JSON
or /only/ text). So we've crafted additional functions that allow
users to express intent of what they'd like to receive.

The following functions `connectJSON`, `connectText`, `connectBLOB`,
etc. allow the client to discriminate on expectation of content from
the server, and error or no-op otherwise.

We still include the original `connect` function for special scenarios where
the server can send both text and binary blobs.

- [x] Add additional `connect*` variants into `Miso.WebSocket`
- [x] Add `textOnly` boolean to conditionally call `JSON.parse`
- [x] Make callbacks nullable so as not to spam Haskell heap and perform unnecessary marshalling logic.

cc @juliendehos