Wire-up `hydrateModel` to `ToHtml` instance.
During server pre-rendering we'd like to invoke hydrateModel. This allows HTML to be generated with desired model state. Since adding Component , hydration is now required for all Component, including nested Component. It can be unwieldy to thread model state through to all sub Component in the ToHtml instance (this is what hydrateModel was introduced to ameliorate), while simultaneously preserving Component encapsulation as enforced by an existential barrier.

This change allows model hydration on the server when prerendering, so the ToHtml Component instance can generate HTML in the same way that the client can generate the virtual DOM, facilitating pointer copying between the two, and hydration.

- [x] Oscillate `JSM` / `IO` when using `jsaddle` or vanilla GHC w/o `jsaddle`
- [x] Invoke `hydrateModel` from within `toHtml` when pre-rendering on the server (w/o `jsaddle`)
- [x] Adds `jsaddle` flag.