Add `VFrag`, `group_` primitive.
This is a primitive and convenience function for inlining children nodes
into a `View m a`. This is meant to avoid situations where it can be
unwieldy to append child node lists together with '++', as shown below.

```haskell
view m =
  div_
  [] $
  [ "foo"
  ] ++
  [ "bar"
  , view1
  ] ++
  [ "baz"
  | shouldInclude
  ]
```

When `group_` is used it inlines all of the nodes into the child
lists.

```haskell
view m =
  div_
  []
  [ group_ [ "foo", "bar", "baz", group_ [ "baz" | shouldInclude ] ]
  ]
``

The `group_` has no DOM representation and is erased when `View m a`
is converted into a JS `VTree<DOMRef>`, with a single exception.

If 'group_' is used a top-level node in a t'Component' it will be
wrapped with a `<div>`. This is a current limitation. So its wise to
only use `group_` in DOM child lists to avoid an additional DOM node being appended.

Note: This is *not* a full implementation of [React
Fragments](https://react.dev/reference/react/Fragment), hence why
the combinator is called 'group_', and only operates at the level
of the Haskell DSL.

React allows components to return child fragments into other components,
while keeping each state separate in its own closure. This means a
child node list could contain nodes from two different components, and
event delegation would need to be adjusted accordingly to account for
this. This is would be a somewhat signifcant change to accomplish in
miso. It would entail adding event listeners to all top-level child
nodes in the fragment.

Because we operate between two heaps (JS and Haskell), and component
state lives in the Haskell heap, it is not as straightforward to implement
Fragments in the same way as React (albeit not impossible).

Further work on React-style fragments is being done by [@ners](https://github.com/ners) in his
[branch](https://github.com/ners/miso).

But this serves as a good base and first approximation for the
feature and solves a major pain point during view templating.

Co-authored-by: ners <ners@gmx.ch>