Drops `shouldSync` property, determine child sync. optimizations in diff dynamically.
The `shouldSync` property was introduced as a preprocessing step to avoid logic in JS
relating to determining if child synch. optimizations should be
entered. Unfortunately, this "optimization" introduced a lot of
additonal allocations on the Haskell side of things.

This is because in order to determine `shouldSync`, we must treat
linked lists as data structures as opposed to control structures,
this breaks streaming into FFI functions for VTree construction in
JSM. This change removes shouldSync logic from VTree.

Co-authored-by: Zer0- <phil.volguine@gmail.com>