Fallback to `setProperty`. (#1231)
This fixes an edge case when using inline styles where CSS variables are not set using
`style[k]=v` syntax, but rather require `node.style.setProperty(k,v)`. Since this
method is slow we only use it as a fallback in the case a custom
variable is in use.

- [x] Update tests to use `backgroundColor`, to avoid jsdom issues
- [x] Regen JS

Confirmed this fixes the slider in `miso-ui`.

cc @ners.