Define `prop` in terms of `ToJSVal`.
This should remove intermediate allocations of `Value` during `View m
a` construction.

- [x] Constrain `prop` on `ToJSVal a => a`
- [x] Define `integerProp` in terms of `doubleProp`.

There is no `ToJSVal` on `Integer`, users can marshall their `Integer`
into a `Double`, or we can add a custom `BigInt` (non-IE
compatible). Previously `integerProp` was marshalled into a
`Scientific` value and then converted into a JS `float`.

Defining `integerProp` in terms of `doubleProp` is identical to this.