Add the `sync` primitive for synchronous I/O execution.
During the implementation of #1262, I noticed async I/O was causing
logs to be printed out of order as the capture / bubble chains were
traversed. Forcing synchronous I/O showed the correct event
propagation ordering.

This has inspired the creation of `sync` primitive that will ensure an
`I/O` is evaluated /before/ the next draw. This should be used
sparingly because it will block the render thread for that specific
`Component`.

```haskell
update = \case
  AddOne -> do
    this += 1
    sync $ consoleLog "added one"
  SubtractOne -> do
    this -= 1
    sync (consoleLog "subtracted one")
```

Note: we still don't allow the evaluation of `I/O` by the user, but
just the scheduling of `I/O` for later evaluation by the `miso` runtime.

cc @brakubraku